Optimize training ~380x: backprop via GradientTape, double Scalar, cached validation
Measured: the NumericalMethods test suite went from 12m42s to 2s. The
slowness was algorithmic; four root causes fixed:

- Function.Evaluate ran the duplicate-parameter check (full BFS + LINQ)
  at every node because recursion went through the public method, making
  evaluation O(n^2) with heavy allocation. The check is now computed once
  during the Params walk and cached (tree structure is immutable), checked
  once at the root, with recursion through a private EvaluateCore.
  Evaluating an unset placeholder now throws naming the parameter.
- New GradientTape implements reverse-mode autodiff (backpropagation):
  topological order and child indexes computed once at construction, then
  each Compute() is two allocation-free array sweeps yielding the value
  and ALL partial derivatives in O(n), replacing P symbolic derivative
  tree evaluations. Function.Backpropagate() is a one-shot convenience
  wrapper. The symbolic Derivative class is untouched and serves as the
  correctness oracle in tests.
- GradientDescend.Train builds the loss (target - f)^2 once, with the
  target as a reserved "__target" placeholder instead of a per-entry
  constant, so substituting a training entry is just placeholder
  assignment: no tree building or re-differentiation inside the loops.
  Both GetMinimun variants use one tape pass per epoch; momentum gained
  the same gradient-threshold early stop as the raw variant.
- Scalar switched from decimal (software-emulated 128-bit) to a readonly
  record struct over double: hardware arithmetic, zero heap allocations
  per operation. Precision was already capped at double by the
  Math.Sin/Exp round-trips.

Also fixes a latent bug: TrainingData kept placeholders in a value-hashed
HashSet whose hash included the mutable Scalar, so training mutations
corrupted lookups; it now uses ReferenceEqualityComparer.

Tests: all suites migrated to double (tolerance asserts where needed);
new FunctionBackpropagationTests pin tape gradients against the symbolic
Derivative oracle to 9 decimal places, including a shared-subtree DAG
case with sin/exp/ln, plus duplicate-placeholder and unset-placeholder
error tests. 98 tests green.

+++ DerivationChainRule.UnitTests/FunctionDerivationTests.cs
@@ -73,9 +73,9 @@ public class FunctionDerivationTests
         for (int xx = 0; xx < 10; xx++)
         {
             xPlaceholder.Scalar = Scalar.Create(xx);
-            var expected = (decimal)(Math.Cos(2 * xx) * 2);
+            var expected = Math.Cos(2 * xx) * 2;
             var actual = df.Evaluate().Value;
-            Assert.True(Math.Abs(expected - actual) < 0.0001m, $"Expected {expected}, got {actual}");
+            Assert.True(Math.Abs(expected - actual) < 0.0001, $"Expected {expected}, got {actual}");
         }
     }
 
@@ -92,9 +92,9 @@ public class FunctionDerivationTests
         for (int xx = 0; xx < 5; xx++)
         {
             xPlaceholder.Scalar = Scalar.Create(xx);
-            var expected = (decimal)(Math.Exp(3 * xx) * 3);
+            var expected = Math.Exp(3 * xx) * 3;
             var actual = df.Evaluate().Value;
-            Assert.True(Math.Abs(expected - actual) < 0.001m, $"Expected {expected}, got {actual}");
+            Assert.True(Math.Abs(expected - actual) < 0.001, $"Expected {expected}, got {actual}");
         }
     }
 
@@ -111,9 +111,9 @@ public class FunctionDerivationTests
         for (int xx = 0; xx < 10; xx++)
         {
             xPlaceholder.Scalar = Scalar.Create(xx);
-            var expected = 1m / (xx + 1);
+            var expected = 1.0 / (xx + 1);
             var actual = df.Evaluate().Value;
-            Assert.True(Math.Abs(expected - actual) < 0.0001m, $"Expected {expected}, got {actual}");
+            Assert.True(Math.Abs(expected - actual) < 0.0001, $"Expected {expected}, got {actual}");
         }
     }
 
@@ -130,9 +130,9 @@ public class FunctionDerivationTests
         for (int xx = 0; xx < 5; xx++)
         {
             xPlaceholder.Scalar = Scalar.Create(xx);
-            var expected = (decimal)(-Math.Sin(4 * xx) * 4);
+            var expected = -Math.Sin(4 * xx) * 4;
             var actual = df.Evaluate().Value;
-            Assert.True(Math.Abs(expected - actual) < 0.0001m, $"Expected {expected}, got {actual}");
+            Assert.True(Math.Abs(expected - actual) < 0.0001, $"Expected {expected}, got {actual}");
         }
     }
 }
